refactor(task-selector): add TaskGrid component
